Rename and move to a new home
