+++ source/Processing.Api/MoveIn/MoveInRequestDto.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Processing.Api.Dtos;
+namespace Processing.Api.MoveIn;
 
 public class MoveInRequestDto
 {
